refactor: 타입별 Nav 구현을 위해 기존 Nav 내부 컴포넌트 추출
